Fixed tx_pool error at startup

+++ core/state/state_object.go
@@ -91,7 +91,7 @@ type stateObject struct {
 	touched   bool
 	deleted   bool
 	onDirty   func(addr common.Address) // Callback method to mark a state object newly dirty
-	idleAge   int
+	idleAge   uint16
 }
 
 // empty returns whether the account is considered empty.
